Adding E2E test case
