+++ apps/teams-test-app/src/App.tsx
@@ -50,6 +50,7 @@ import SecondaryBrowserAPIs from './components/SecondaryBrowserAPIs';
 import SharingAPIs from './components/SharingAPIs';
 import StageViewAPIs from './components/StageViewAPIs';
 import TeamsCoreAPIs from './components/TeamsCoreAPIs';
+import ThirdPartyStorageProvidersAPIs from './components/ThirdPartyStorageProvidersAPIs';
 import { isTestBackCompat } from './components/utils/isTestBackCompat';
 import Version from './components/Version';
 import VideoAPIs from './components/VideoEffectsApis';
@@ -122,6 +123,7 @@ const App = (): ReactElement => {
   return (
     <div>
       <div className="App-container">
+        <ThirdPartyStorageProvidersAPIs />
         <AppAPIs />
         <AppInitializationAPIs />
         <AppInstallDialogAPIs />
